feature/1.0.0/Base/App/Study/DataBinding/moveFiles_Models_Views #25
- DataBindingView.swift
- DataBindingModel.swift

TODO
- [ ] DataBindingVariable.swift
